// matthews-case-management/src/app/models/case-to-facility-status.model.ts
export class CaseToFacilityStatus {
    caseId: string = '00000000-0000-0000-0000-000000000000';
    facilityStatusId: string = '00000000-0000-0000-0000-000000000000';
    isDone: boolean = false;
    createdBy: string = '00000000-0000-0000-0000-000000000000';
    createdTime: string = '0001-01-01T00:00:00';
    ModifiedBy: string = '00000000-0000-0000-0000-000000000000';
    ModifiedTime: string = '0001-01-01T00:00:00';
}